refactor(store): Move user command key helpers to separate mod
