+++ rust/src/utility/store/command/mod.rs
@@ -1,1 +1,2 @@
 pub mod internal;
+pub mod user;
